move functionality to separate module
